requirements.txt
requests~=2.27.1
python-dotenv~=0.20.0
python-telegram-bot-calendar~=1.0.5
aiogram~=2.21
peewee~=3.15.1
psycopg2-binary~=2.9.3